+++ src/utils/http.js
@@ -1,5 +1,6 @@
 const axios = require('axios')
 const joi = require('@hapi/joi')
+const i18nUtils = require('./i18n')
 
 const schema = joi.object().keys({
   method: joi.string(),
@@ -63,7 +64,27 @@ const request = (params) => {
   })
 }
 
+const error = (res, message, status) => {
+  console.log(message || {}, status || {})
+  res
+    .status(status || 500)
+    .set('Content-Type', 'application/json')
+    .json({
+      status: status || 500,
+      message: (message || i18nUtils.translate('generic_error'))
+    })
+}
+
+const json = (res, data) => {
+  res
+    .status(200)
+    .set('Content-Type', 'application/json')
+    .json(data)
+}
+
 module.exports = {
+  error,
+  json,
   post,
   get
 }
